fixed: team_stat.rb name
